requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.0
defusedxml==0.8.0rc2
Django==4.2.15
django-environ==0.11.2
djangorestframework==3.15.2
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-pubsub==2.23.0
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
grpc-google-iam-v1==0.13.1
grpcio==1.65.4
grpcio-status==1.65.4
httplib2==0.22.0
idna==3.7
oauthlib==3.2.2
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyotp==2.9.0
pyparsing==3.1.2
pypng==0.20220715.0
python3-openid==3.2.0
qrcode==7.4.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
